chore: move getting feature flags to own file and out of fetch since it's not async
