+++ packages/bridge-controller/src/utils/fetch.test.ts
@@ -1,8 +1,8 @@
 import { AddressZero } from '@ethersproject/constants';
 import type { CaipAssetType } from '@metamask/utils';
 
+import { getBridgeFeatureFlags } from './feature-flags';
 import {
-  fetchBridgeFeatureFlags,
   fetchBridgeQuotes,
   fetchBridgeTokens,
   fetchAssetPrices,
@@ -10,223 +10,10 @@ import {
 import mockBridgeQuotesErc20Erc20 from '../../tests/mock-quotes-erc20-erc20.json';
 import mockBridgeQuotesNativeErc20 from '../../tests/mock-quotes-native-erc20.json';
 import { BridgeClientId, BRIDGE_PROD_API_BASE_URL } from '../constants/bridge';
-import type { BridgeControllerMessenger } from '../types';
 
 const mockFetchFn = jest.fn();
 
-const mockMessenger = {
-  call: jest.fn(),
-  publish: jest.fn(),
-  registerActionHandler: jest.fn(),
-  registerInitialEventPayload: jest.fn(),
-} as unknown as BridgeControllerMessenger;
-
 describe('fetch', () => {
-  describe('fetchBridgeFeatureFlags', () => {
-    it('should fetch bridge feature flags successfully', async () => {
-      const bridgeConfig = {
-        refreshRate: 3,
-        maxRefreshCount: 1,
-        support: true,
-        chains: {
-          '1': {
-            isActiveSrc: true,
-            isActiveDest: true,
-          },
-          '10': {
-            isActiveSrc: true,
-            isActiveDest: false,
-          },
-          '59144': {
-            isActiveSrc: true,
-            isActiveDest: true,
-          },
-          '120': {
-            isActiveSrc: true,
-            isActiveDest: false,
-          },
-          '137': {
-            isActiveSrc: false,
-            isActiveDest: true,
-          },
-          '11111': {
-            isActiveSrc: false,
-            isActiveDest: true,
-          },
-          '1151111081099710': {
-            isActiveSrc: true,
-            isActiveDest: true,
-          },
-        },
-      };
-
-      const remoteFeatureFlagControllerState = {
-        cacheTimestamp: 1745515389440,
-        remoteFeatureFlags: {
-          bridgeConfig,
-          assetsNotificationsEnabled: false,
-          confirmation_redesign: {
-            contract_interaction: false,
-            signatures: false,
-            staking_confirmations: false,
-          },
-          confirmations_eip_7702: {},
-          earnFeatureFlagTemplate: {
-            enabled: false,
-            minimumVersion: '0.0.0',
-          },
-          earnPooledStakingEnabled: {
-            enabled: false,
-            minimumVersion: '0.0.0',
-          },
-          earnPooledStakingServiceInterruptionBannerEnabled: {
-            enabled: false,
-            minimumVersion: '0.0.0',
-          },
-          earnStablecoinLendingEnabled: {
-            enabled: false,
-            minimumVersion: '0.0.0',
-          },
-          earnStablecoinLendingServiceInterruptionBannerEnabled: {
-            enabled: false,
-            minimumVersion: '0.0.0',
-          },
-          mobileMinimumVersions: {
-            androidMinimumAPIVersion: 0,
-            appMinimumBuild: 0,
-            appleMinimumOS: 0,
-          },
-          productSafetyDappScanning: false,
-          testFlagForThreshold: {},
-          tokenSearchDiscoveryEnabled: false,
-          transactionsPrivacyPolicyUpdate: 'no_update',
-          transactionsTxHashInAnalytics: false,
-          walletFrameworkRpcFailoverEnabled: false,
-        },
-      };
-
-      (mockMessenger.call as jest.Mock).mockImplementation(() => {
-        return remoteFeatureFlagControllerState;
-      });
-
-      const result = await fetchBridgeFeatureFlags(mockMessenger);
-
-      const expectedBridgeConfig = {
-        maxRefreshCount: 1,
-        refreshRate: 3,
-        support: true,
-        chains: {
-          'eip155:1': {
-            isActiveDest: true,
-            isActiveSrc: true,
-          },
-          'eip155:10': {
-            isActiveDest: false,
-            isActiveSrc: true,
-          },
-          'eip155:11111': {
-            isActiveDest: true,
-            isActiveSrc: false,
-          },
-          'eip155:120': {
-            isActiveDest: false,
-            isActiveSrc: true,
-          },
-          'eip155:137': {
-            isActiveDest: true,
-            isActiveSrc: false,
-          },
-          'eip155:59144': {
-            isActiveDest: true,
-            isActiveSrc: true,
-          },
-          'solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp': {
-            isActiveDest: true,
-            isActiveSrc: true,
-          },
-        },
-      };
-
-      expect(result).toStrictEqual(expectedBridgeConfig);
-    });
-
-    it('should use fallback bridge feature flags if response is unexpected', async () => {
-      const bridgeConfig = {
-        refreshRate: 3,
-        maxRefreshCount: 1,
-        support: 25,
-        chains: {
-          a: {
-            isActiveSrc: 1,
-            isActiveDest: 'test',
-          },
-          '2': {
-            isActiveSrc: 'test',
-            isActiveDest: 2,
-          },
-        },
-      };
-      const remoteFeatureFlagControllerState = {
-        cacheTimestamp: 1745515389440,
-        remoteFeatureFlags: {
-          bridgeConfig,
-          assetsNotificationsEnabled: false,
-          confirmation_redesign: {
-            contract_interaction: false,
-            signatures: false,
-            staking_confirmations: false,
-          },
-          confirmations_eip_7702: {},
-          earnFeatureFlagTemplate: {
-            enabled: false,
-            minimumVersion: '0.0.0',
-          },
-          earnPooledStakingEnabled: {
-            enabled: false,
-            minimumVersion: '0.0.0',
-          },
-          earnPooledStakingServiceInterruptionBannerEnabled: {
-            enabled: false,
-            minimumVersion: '0.0.0',
-          },
-          earnStablecoinLendingEnabled: {
-            enabled: false,
-            minimumVersion: '0.0.0',
-          },
-          earnStablecoinLendingServiceInterruptionBannerEnabled: {
-            enabled: false,
-            minimumVersion: '0.0.0',
-          },
-          mobileMinimumVersions: {
-            androidMinimumAPIVersion: 0,
-            appMinimumBuild: 0,
-            appleMinimumOS: 0,
-          },
-          productSafetyDappScanning: false,
-          testFlagForThreshold: {},
-          tokenSearchDiscoveryEnabled: false,
-          transactionsPrivacyPolicyUpdate: 'no_update',
-          transactionsTxHashInAnalytics: false,
-          walletFrameworkRpcFailoverEnabled: false,
-        },
-      };
-
-      (mockMessenger.call as jest.Mock).mockResolvedValue(
-        remoteFeatureFlagControllerState,
-      );
-
-      const result = await fetchBridgeFeatureFlags(mockMessenger);
-
-      const expectedBridgeConfig = {
-        maxRefreshCount: 5,
-        refreshRate: 30000,
-        support: false,
-        chains: {},
-      };
-      expect(result).toStrictEqual(expectedBridgeConfig);
-    });
-  });
-
   describe('fetchBridgeTokens', () => {
     it('should fetch bridge tokens successfully', async () => {
       const mockResponse = [
